### 2020-08-31 #2619 @whedon
[PRE REVIEW]: SCALAR - A Platform for Real-time Machine Learning Competitions on Data Streams
**Submitting author:** @nedRad88 (<a href="http://orcid.org/0000-0001-5056-6270">Nedeljko Radulovic</a>)
**Repository:** <a href="https://github.com/nedRad88/SCALAR" target ="_blank">https://github.com/nedRad88/SCALAR</a>
**Version:** v0.1
