### 2020-09-04 #2619 @galessiorob
👋 Hi @nedRad88

I am in the process of assigning reviewers to this submission, but if you have any suggestions for reviewers please let me know.

In the meantime, it looks like you're only missing 1 DOIs, if you could nail that one down that would be great.

Many thanks!

Basic checks:

 - [ ] There’s a general description of the software
 - [ ] The software is within the scope as research software
 - [ ] OSI-approved license: GNU General Public License (GPL)
