+++ provider/paper/references.bib
@@ -37,6 +37,7 @@
   booktitle={2019 IEEE International Conference on Big Data (Big Data)},
   pages={3493--3497},
   year={2019},
+  doi = {10.1109/BigData47090.2019.9006357},
   organization={IEEE}
 }
 
