### 2020-09-15 #2619 @galessiorob
@atanikan & @GregaVrbancic Thank you so much for being reviewers for this paper!

I will close this issue and generate a new one that will contain individual checklists for you to work through. Please let me know if something is not clear or if you need help in any way. Looking forward to collaborating with you! 🎉
