### 2020-09-15 #2676 @whedon
[REVIEW]: SCALAR - A Platform for Real-time Machine Learning Competitions on Data Streams
**Submitting author:** @nedRad88 (<a href="http://orcid.org/0000-0001-5056-6270">Nedeljko Radulovic</a>)
**Repository:** <a href="https://github.com/nedRad88/SCALAR" target ="_blank">https://github.com/nedRad88/SCALAR</a>
**Version:** v0.1
