### 2020-09-23 nedRad88/SCALAR#3 @GregaVrbancic
[JOSS Review] Installation instructions
Hello @nedRad88, I am the reviewer of your SCALAR paper on JOSS  openjournals/joss-reviews/issues/2676. While there are installation instructions available utilizing Docker and docker-compose technology, I think it would be beneficial to also include the python dependencies list in the documentation. Ideally, the python dependecy should be maintained using package management tools such as pip, poetry, pipenv.
