### 2020-09-23 nedRad88/SCALAR#4 @GregaVrbancic
[JOSS Review] - Test functionality
Hi @nedRad88, as required by openjournals/joss-reviews/issues/2676 checklist could you provide some automatic tests or manual steps, so that the functionality of the software can be verified.
