### 2020-09-23 nedRad88/SCALAR#5 @GregaVrbancic
[JOSS Review] - Documentation
Hi @nedRad88, could you provide some documentation for main API functionalities as required by openjournals/joss-reviews/issues/2676 revision?
