### 2020-09-23 nedRad88/SCALAR#6 @GregaVrbancic
[JOSS Review] - Community guidelines
Hi @nedRad88, could you please provide community guidelines, how the community could contribute to this project. This issue references to the review issue openjournals/joss-reviews/issues/2676.

Here are some examples of community guidelines of popular open-source projects:

- https://github.com/atom/atom/blob/master/CONTRIBUTING.md
- https://github.com/rails/rails/blob/master/CONTRIBUTING.md
- https://github.com/opengovernment/opengovernment/blob/master/CONTRIBUTING.md
