### 2020-09-23 #2676 @GregaVrbancic
Hi @galessiorob and @nedRad88,
I have made some progress on this review. I have opened several issues on SCALAR repository, so @nedRad88, please take a look at them. Also, please clarify Dihia Boulegane's and Albert Bifet's contributions to the paper since it is not observable from commit history (which can happen sometimes). Additionally, based on the commit history, should @gradoslovar be included as an co-author?
