### 2020-09-23 01ea819
Updated the documentation and added contribution guidelines
