+++ provider/my_application/views.py
@@ -14,7 +14,6 @@ See the License for the specific language governing permissions and
 limitations under the License.
 """
 
-
 from __future__ import print_function
 import time
 time.sleep(15)
@@ -37,7 +36,6 @@ from repository import MongoRepository
 import logging
 from itsdangerous import URLSafeTimedSerializer
 import csv
-
 from io import StringIO
 from werkzeug.datastructures import Headers
 from hashids import Hashids
@@ -137,32 +135,29 @@ _MONGO_REPO.insert_standard_measures(standard_measures=standard_measures)
 
 
 def authorized(*roles):
+    """
+    Give authorization only to ADMIN user for some routes.
+
+    :param roles:
+    :return:
+    """
     def wrapper(f):
         @wraps(f)
         def wrapped(*args, **kwargs):
-
             token = request.headers.get('Authorization')
-
             decoded = decode_auth_token(token)
             # print decoded
             if decoded[0] != 200:
                 return decoded[1]
-
             user = decoded[1]
             logging.debug("Decorator: {}".format(user))
-
             if roles:
-
                 user_roles = [str(r) for r in user['roles']]
-
                 if "roles" not in user or len(set(user_roles).intersection(set(roles))) == 0:
                     # return "Not Allowed"
                     return json.dumps('Forbidden Admin Only'), 403, {'ContentType': 'application/json'}
-
             return f(user, *args, **kwargs)
-
         return wrapped
-
     return wrapper
 
 
@@ -178,6 +173,11 @@ def index():
 # TODO: add checking on email confirmation and add decorator
 @app.route('/auth/login', methods=['POST'])
 def login():
+    """
+    Login route. Provide the credentials (email/password).
+
+    :return:
+    """
     user_name = request.form['username']
     password = request.form['password']
 
@@ -201,6 +201,12 @@ def me(user):
 
 @app.route('/auth/api/account/confirm/<token>')
 def confirm_email(token):
+    """
+    Check if the confirmation token is correct.
+
+    :param token:
+    :return:
+    """
     logging.debug("Token: {}".format(token))
     try:
         email = confirm_token(token)
@@ -223,7 +229,15 @@ def confirm_email(token):
 
 @app.route('/auth/register', methods=['POST'])
 def register():
+    """
+    Registration route. Provides a form to register to the platform.
+        Parameters:
+            - First name, last name, email, password
+    Then the confirmation e-mail is sent with a token to validate the registration.
 
+    :return: Response:
+                - 200: confirm success
+    """
     data = json.loads(request.data.decode('utf-8'))
     web = config["WEB_ADDRESS"]
     first_name = data['firstName']
@@ -244,7 +258,6 @@ def register():
     msg.body = "Hello " + first_name + ' ' + last_name + ", \n\nWelcome to Streaming Data Challenge platform! \n\nCheers, \nThe team \n\nPlease click on the link below to confirm your e-mail.\n" "http://" + web + ":80/auth/api/account/confirm/" + token
 
     mail.send(msg)
-
     return json.dumps({'success': True}), 200, {'ContentType': 'application/json'}
 
 
@@ -260,7 +273,6 @@ def get_users(user):
 def delete_users(user):
     users = json.loads(request.data)
     _USER_REPO.delete_many(users)
-
     return json.dumps({'success': True}), 200, {'ContentType': 'application/json'}
 
 
@@ -276,6 +288,24 @@ def test():
 
 @app.route('/api/competitions', methods=['POST', 'GET'])
 def competitions():
+    """
+    Competitions route.
+    Implements two methods:
+    GET: retrieve the competitions
+    POST: create new competition:
+        Parameters:
+            - Competition name
+            - Datastream
+            - Description of the competition
+            - Competition settings: size of initial batch, initial training time, size of a regular batch,
[…]

[…]
