### 2020-09-24 6762e95
updated contributing and functionality description
