+++ README.md
@@ -5,6 +5,35 @@
 SCALAR is the first of its kind, platform to organize Machine Learning competitions on Data Streams.
 It was used to organize a real-time ML competition on IEEE Big Data Cup Challenges 2019.
 
+## Features
+
+**Data stream mining competitions** - With SCALAR you can organize the real-time competitions on Data Streams. 
+It is inspired by [Kaggle](https://www.kaggle.com/), platform for offline machine learning competitions.
+
+
+
+**Simple, user friendly interface** - SCALAR has a WEB application that allows you to easily browse and 
+subscribe to the competitions. Its simple and intuitive design, let's you to easily upload the datasets, and create the 
+competition. 
+
+![Dialog competition](./add_competition.png)
+
+**Live results and leaderboard** - Since the competition is real-time, the results are also updated in real time. 
+During the competition, you can follow performance of your model in the WEB application. Leaderboard will show how 
+do you compare to other users, and live chart shows the comparison with baseline model.
+
+**Secure, bi-directional streaming communication** - We use a combination of `gRPC` and `Protobuf` to provide secure, 
+low latency bi-directional streaming communication between server and users.
+
+![Architecture](./Architecture.png)
+
+**Freedom to choose a programming language** - SCALAR let's users to choose their preferred environment. The only 
+requirement is to be able to communicate through `gRPC` and `Protobuf`, which is supported for many programming 
+languages: Python, Java, C++, GO... Additionally, SCALAR provides support for R. Apart from that, users can choose 
+their setup, environment and additional resources to train better models.
+
+**** 
+
 ## Getting Started
 
 The project is done in Python and organized in Docker containers.
@@ -87,3 +116,5 @@ Open source Docker containers were used:
 
 * [Boulegane, Dihia, et al. "Real-Time Machine Learning Competition on Data Streams at the IEEE Big Data 2019." 2019 IEEE International Conference on Big Data (Big Data). IEEE, 2019.](https://ieeexplore.ieee.org/abstract/document/9006357?casa_token=f0mJeR8-WfYAAAAA:yEt_Mix9dumrPpo64uPBbI0XI4Kvfim4Pkg5xNVVzXqK4AGToX0XcJPKgETkE1hs86Pcc0u5xYc)
 * [IEEE Big Data Cup 2019](https://bigmine.github.io/real-time-ML-competition/index.html)
+
+
