+++ pull_request_template.md
@@ -1,0 +1,27 @@
+<!-- 
+Thank you for contributing with a PR!
+
+Please fill the description of change(s) and/or indicate if it fixes an open issue (optional).
+
+To ease the merge process please review the attached checklist.
+-->
+
+Changes proposed in this pull request:
+
+* 
+* 
+* Fixes # [Optional]
+
+---
+
+### Checklist
+#### Implementation
+- [ ] Implementation is correct (it performs its intended function).
+- [ ] Code is consistent with the framework.
+- [ ] Code is properly documented.
+- [ ] PR description covers ALL the changes performed.
+- [ ] Files changed (update, add, delete) are in the PR's scope (no extra files are included).
+
+#### Tests
+- [ ] New functionality is tested.
+- [ ] ALL tests pass with no errors.
