+++ CONTRIBUTING.md
@@ -13,6 +13,9 @@ If you are not familiar with GitHub and how to contribute to a github project, c
  - [The Beginner's guide to contributing to a GitHub project](https://akrabat.com/the-beginners-guide-to-contributing-to-a-github-project/)
 
  - [Make a clean pull request](https://github.com/MarcDiethelm/contributing/blob/master/README.md)
+ 
+If you have an idea what can be improved or added to the system, or you found a bug report an [issue](https://github.com/nedRad88/SCALAR/issues).
+We will respond promptly to discuss the
 
 ## Code style and documentation
 
