+++ example_data/test.md
@@ -1,0 +1,6 @@
+### Test procedure
+
+To test the functionality of the platform we provide you the example dataset and the list of manual steps to perform.
+Complete the installation (following the instructions), and once you have registered yourself as an `ADMIN` user,
+proceed with the following steps to test the software:
+
