### 2020-09-25 #2676 @nedRad88
Hi @GregaVrbancic,
Thanks for the remarks.
Regarding authorship, the repository was not established back then when the project was started, so that is why Albert and Dihia don't appear in the commit history. If you check the second commit (https://github.com/nedRad88/SCALAR/tree/944fe1ae3eda7e3b2d481f81618dfd71d38cfa92) you can see that the project existed before creating the repository.
Regarding @gradoslovar, he is an external contributor, but yes we can add him as a co-author. Should I update the paper?
Thanks, I will come back to you soon to resolve all other the opened issues.
