### 2020-09-28 nedRad88/SCALAR#3 @xiaohk
Hello @nedRad88, I am another reviewer for your JOSS submission.

I am following the installation instruction on README, and have some problem to run the platform in my browser. I am currently stuck at the step to install `sql_db` container. I also wonder how to run the app after all these dependences are installed; can I just run `docker run`?

The installation instruction would be much better if it lists _all exact commands_ that users need to run, in order to open this web app in their browsers.
