### 2020-09-28 nedRad88/SCALAR#5 @xiaohk
I found a [documentation](https://github.com/nedRad88/SCALAR/blob/master/SCALAR_Quick_Start_Guide.pdf) for the competition participants.
