### 2020-09-28 #2676 @xiaohk
Hello @nedRad88. Very helpful and interesting work!

I have added some comments on the issues raised by @GregaVrbancic on SCALAR repo.

My main concern is that the installation instruction is not very easy to follow. It would be much easier for other users to set up their SCALAR if the instruction provides a list of _all exact commands_ that they needs to run in order to set it up.

Other minor comments:

- Either on the README page or on the [documentation](), I suggest adding some text to discuss the problem and motivation for this software. I found the text on [Big Data Challenge](https://bigmine.github.io/real-time-ML-competition/index.html) very helpful.
- It would be great to have some automated tests and community guidelines
- A typo on the README page: SCALAR `let's` users → `lets`
