### 2020-09-29 nedRad88/SCALAR#3 @nedRad88
Hello @xiaohk,
Thanks for the comments.
Did you manage to finish the installation? If not can you tell me what is the problem that you face? After installing `Docker`, `Docker compose` and setting `config.json` and `docker-compose.yml` according to instructions, you run `docker-compose up` and it will pull all necessary containers and run them. I will make the instructions more clear and come back to you.
