### 2020-09-29 6922003
added @gradoslovar to acknowlegments and updated the contribution instructions
