### 2020-09-29 a71038f
typos in readme
