### 2020-09-30 nedRad88/SCALAR#3 @xiaohk
Thanks for your promptly reply.

Can you please also provide some instruction for installing `sql_db`? Thanks!
