### 2020-10-01 856c92c
updated the installation instructions and requirements
