+++ provider/my_application/requirements.txt
@@ -1,24 +1,26 @@
-kafka-python==1.4.2
-python-dateutil==2.7.0
-apscheduler==3.5.1
-sqlalchemy==1.2.5
-pymysql==0.8.0
-Flask==0.12.2
-Flask-SQLAlchemy==2.3.2
-Flask-WTF==0.14.2
-werkzeug==0.14.1
-grpcio==1.10.0
-grpcio-tools==1.10.0
-multiprocessing==2.6.2.1
-pymongo==3.6.1
-flask-socketio==2.9.6
-eventlet==0.22.1
-six==1.10.0
-redis==2.10.6
-flask_sse==0.2.1
-pyjwt==1.6.1
-Flask-Mail==0.9.1
-greenlet==0.4.13
-gevent==1.2.2
-pause==0.1.2
-hashids==1.2.0
+kafka-python
+python-dateutil
+apscheduler
+sqlalchemy
+pymysql
+Flask
+Flask-SQLAlchemy
+Flask-WTF
+werkzeug
+grpcio==1.13.0 --no-binary grpcio
+grpcio-tools==1.13.0
+pymongo
+flask-socketio
+eventlet
+redis
+flask_sse
+pyjwt
+Flask-Mail
+greenlet
+gevent
+pause
+hashids
+pyspark==2.4.0
+confluent-kafka
+orjson
+scikit-multiflow
