### 2020-10-01 nedRad88/SCALAR#3 @nedRad88
Hello @xiaohk,
Did you mean for installing mysql-client? We are working on improving the installation instructions and now this step has been removed. Now, `docker-compose up` will start the platform. You just need the latest version of the code.
