### 2020-10-01 nedRad88/SCALAR#6 @nedRad88
Hi @GregaVrbancic,

the repository has been updated with [contributing.md](https://github.com/nedRad88/SCALAR/blob/master/CONTRIBUTING.md) and the [pull request template](https://github.com/nedRad88/SCALAR/blob/master/pull_request_template.md).
