### 2020-10-01 nedRad88/SCALAR#5 @nedRad88
Hi @GregaVrbancic,

The documentation has been updated, and also [README.md](https://github.com/nedRad88/SCALAR/blob/master/README.md), saying more about the main API functionalities.
@xiaohk What you sharedm it is a [Tutorial](https://github.com/nedRad88/SCALAR/blob/master/SCALAR_Quick_Start_Guide.pdf) on how to use a platform. It provides the information how to set up the competition but also how to participate in one. We will soon provide a simpler procedure to do this and test the functionality of the platform.
