### 2020-10-01 #2676 @nedRad88
Hello @galessiorob , @xiaohk ,
Thanks for the comments.
I have added @gradoslovar as a contributor.
Regarding other issues, we are working on them, I am posting updates on each issue, separately.
We will provide the updated procedure to test the software soon and we are simplifying the installation procedure.
