### 2020-10-01 nedRad88/SCALAR#7 @atanikan
[Joss Review] - Potentially create a bloated SCALAR platform Docker image for testing and/or easing installation
https://github.com/openjournals/joss-reviews/issues/2676

Hi, I am one of the reviewers of the journal. I am currently going through your setup documentation. I was wondering
is it even possible to create a bloated up image with all the dependencies pre-set as mentioned in your docker-compose.yml and config file, pre-installed into a single image. I understand there are some interactions needed with the host machine, but maybe if the necessary folders are created inside the image and not mounted, the user can potentially copy the folder from the image onto host machine if and when needed.

 Alternatively, One other suggestion is to have a shell script or python script which guides users in generating the docker-compose.yml and/or the config file based on their input. For e.g. if the user runs

```
python setup.py
Did you install mysql (y/n)? y
Please set your Admin role. Press y to continue
Please mention the hostname?
Please mention the directory/volume on your host machine to be mounted to the docker image. Press enter to use default (/usr/local/bin)?
```
