### 2020-10-11 #2676 @galessiorob
Hey @nedRad88 mind commenting on this issue which of the missing items have you added or fixed so the reviewers and I can go back and check? Eg. Tests, statement of purpose, etc.

Thanks!
