### 2020-10-12 nedRad88/SCALAR#5 @GregaVrbancic
Hi @nedRad88, could you provide me a link to the api documentation? Inam having trouble to find it.
