### 2020-10-12 nedRad88/SCALAR#6 @GregaVrbancic
Hi @nedRad88, this looks good to me.
