### 2020-10-14 #2676 @nedRad88
Hey @galessiorob , sorry for the delay. We have updated the community guidelines and installation instructions. Documentation and improved instructions for tests are under way. I will be posting updates in issue's thread. Thanks!
