+++ provider/my_application/views.py
@@ -1,18 +1,17 @@
-"""
-Copyright 2020 Nedeljko Radulovic, Dihia Boulegane, Albert Bifet
 
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-"""
+# Copyright 2020 Nedeljko Radulovic, Dihia Boulegane, Albert Bifet
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 
 from __future__ import print_function
 import time
@@ -245,12 +244,11 @@ def confirm_email(token):
 def register():
     """
     Registration route. Provides a form to register to the platform.
-        Parameters:
-            - First name, last name, email, password
+    Parameters:
+    - First name, last name, email, password
     Then the confirmation e-mail is sent with a token to validate the registration.
 
-    :return: Response:
-                - 200: confirm success
+    :return: Response: {200: confirm success}
     """
     data = json.loads(request.data.decode('utf-8'))
     web = config["WEB_ADDRESS"]
@@ -307,18 +305,15 @@ def competitions():
     Implements two methods:
     GET: retrieve the competitions
     POST: create new competition:
-        Parameters:
-            - Competition name
-            - Datastream
-            - Description of the competition
-            - Competition settings: size of initial batch, initial training time, size of a regular batch,
-                                    time interval between the batches, name of the label column and the evaluation
-                                    metric, start and end date, time interval to send the predictions, .proto file
+    Parameters:
+    - Competition name
+    - Datastream
+    - Description of the competition
+    - Competition settings: size of initial batch, initial training time, size of a regular batch,
+    time interval between the batches, name of the label column and the evaluation
+    metric, start and end date, time interval to send the predictions, .proto file
 
-    :return: Responses:
-            - 200: If method== "GET", return the list of the competitions;
-                   If method == "POST", confirm success
-            - 500: Error
+    :return: Responses: {200: If method== "GET", return the list of the competitions, If method == "POST", confirm success, 500: Error}
     """
     if request.method == 'GET':
         status = request.args.get('status')
@@ -429,16 +424,12 @@ def get_datastreams():
     Datastreams route. It implements 2 methods:
     - GET: retrieve already existing datasets
     - POST: add new dataset
-        Parameters:
-                - Dataset name
-                - Dataset description
-                - Dataset file
+    Parameters:
+    - Dataset name
+    - Dataset description
+    - Dataset file
 
-    :return: Responses:
-                - If method == "GET": return the list of the datasets
-                - If method == "POST":
-                    - 200: confirm success
-                    - 500: Error
+    :return: Responses: {If method == "GET": return the list of the datasets, If method == "POST": {200: confirm success, 500: Error}
     """
     if request.method == 'GET':
[…]
