### 2020-10-16 fc95832
Set theme jekyll-theme-slate
