### 2020-10-19 nedRad88/SCALAR#5 @nedRad88
Hi @GregaVrbancic, the documentation is available at https://nedrad88.github.io/scalarDocs.github.io/
I will update the installation instructions next and update the information about the documentation on the repository.
