### 2020-10-20 nedRad88/SCALAR#5 @GregaVrbancic
@nedRad88 Thanks. Great work on documentation!
