### 2020-10-23 a80aaa2
updated installing instructions
