### 2020-10-23 4a1bc4b
setup.py small fix
