### 2020-10-23 nedRad88/SCALAR#3 @nedRad88
Hello @GregaVrbancic @xiaohk , I have updated the installation instructions. There is an option for quick setup when you want just to check the functionality of the platform. According to this, I will create a simplified scenario to test the functionality.
