### 2020-10-30 e7afec3
Minor syntax suggestions
