### 2020-10-30 #2676 @galessiorob
@nedRad88 I added a few edit suggestions to the text on the `patch-1` branch, feel free to further edit them if you want.

Looking forward to taking this paper across the finish line!
