### 2020-11-02 8ec61e8
enabling automatic tests
