### 2020-11-02 #2676 @nedRad88
Hi @galessiorob , thanks for the suggestions. I have merged them into master. I am finishing the test setup. I expect to have it done in a day or two. Thank you!
