+++ provider/my_application/views.py
@@ -42,7 +42,7 @@ from hashids import Hashids
 import eventlet
 from random import randint
 eventlet.monkey_patch(time=True)
-logging.basicConfig(level='DEBUG')
+logging.basicConfig(level='INFO')
 
 with open('config.json') as json_data_file:
     config = json.load(json_data_file)
@@ -155,6 +155,12 @@ if _DATASTREAM_REPO.get_datastream_by_id(1) is None:
     _DATASTREAM_REPO.session.commit()
 ############################
 "Create a test competition"
+create_test_competition = True
+competitions_to_come = _COMPETITION_REPO.session.query(Competition).filter(Competition.start_date >= datetime.now()).all()
+logging.debug("COMPETITION TO COME!!!: {}".format(competitions_to_come))
+
+if len(competitions_to_come) > 0:
+    create_test_competition = False
 
 
 def format_time(t):
@@ -162,47 +168,48 @@ def format_time(t):
     return s[:-7]
 
 
-now = datetime.now() + timedelta(seconds=60)
-end_date = datetime.now() + timedelta(minutes=30)
-name = "test" + str(randint(0, 10000))
-generated_code_directory = os.path.join(config['UPLOAD_REPO'], config['COMPETITION_GENERATED_CODE'], name)
-proto_directory = os.path.join(config['UPLOAD_REPO'], config['COMPETITION_PROTO_REPO'], name)
+if create_test_competition:
+    now = datetime.now() + timedelta(seconds=300)
+    end_date = datetime.now() + timedelta(minutes=30)
+    name = "test" + str(randint(0, 10000))
+    generated_code_directory = os.path.join(config['UPLOAD_REPO'], config['COMPETITION_GENERATED_CODE'], name)
+    proto_directory = os.path.join(config['UPLOAD_REPO'], config['COMPETITION_PROTO_REPO'], name)
 
-if not os.path.exists(proto_directory):
-    os.makedirs(proto_directory)
+    if not os.path.exists(proto_directory):
+        os.makedirs(proto_directory)
 
-copy("file.proto", proto_directory)
+    copy("file.proto", proto_directory)
 
-if not os.path.exists(generated_code_directory):
-    os.makedirs(generated_code_directory)
-try:
-    with open(generated_code_directory + '/__init__.py', "w+") as f:
-        f.write('')
-except Exception:
-    pass
+    if not os.path.exists(generated_code_directory):
+        os.makedirs(generated_code_directory)
+    try:
+        with open(generated_code_directory + '/__init__.py', "w+") as f:
+            f.write('')
+    except Exception:
+        pass
 
-try:
-    protoc.main(('', '-I' + proto_directory, '--python_out=' + generated_code_directory,
-                 '--grpc_python_out=' + generated_code_directory,
-                 os.path.join(proto_directory, 'file.proto')))
-except Exception as e:
-    print(str(e))
-competition_config = {"target": ["MAPE"]}
-code = ''
-competition = Competition(None, name=name, datastream_id=1, initial_batch_size=20,
-                                  initial_training_time=20, batch_size=5,
-                                  time_interval=5, target_class="target", start_date=format_time(now),
-                                  file_path="file.proto", predictions_time_interval=5,
-                                  end_date=format_time(end_date), description="Functionality test competition", code=code)
+    try:
+        protoc.main(('', '-I' + proto_directory, '--python_out=' + generated_code_directory,
+                    '--grpc_python_out=' + generated_code_directory,
+                    os.path.join(proto_directory, 'file.proto')))
+    except Exception as e:
+        print(str(e))
+    competition_config = {"target": ["MAPE"]}
+    code = ''
+    competition = Competition(None, name=name, datastream_id=1, initial_batch_size=20,
+                                    initial_training_time=20, batch_size=5,
[…]
