### 2020-11-02 nedRad88/SCALAR#7 @nedRad88
Hi @atanikan , I have created something like you suggested. Please check the new README file and follow new installation instructions to test the functionality of the software. Thanks for this suggestion.
