### 2020-11-02 nedRad88/SCALAR#4 @nedRad88
Hi @GregaVrbancic , @xiaohk , I have followed the suggestion from @atanikan and created a kind of bloated version of the software so the testing the functionality is much easier now. I have updated installation instructions and README file. So, please follow the new instructions to install and run the test competition using our software. Thanks!
