### 2020-11-13 nedRad88/SCALAR#4 @galessiorob
👋 @GregaVrbancic @xiaohk

Checking in on this requirement, could you please go over the expanded testing and install instructions and close this issue if you two are satisfied with the changes?

Thanks!
