### 2020-11-13 #2676 @galessiorob
@nedRad88 the paper looks almost ready to ship, I found a missing space that I added [here.](https://github.com/nedRad88/SCALAR/pull/9)

@GregaVrbancic, @atanikan, @xiaohk could you go over all the added changes and check the missing items from your lists if you're satisfied with them?

Thank you all, we are almost there!
