+++ provider/paper/paper.md
@@ -44,7 +44,7 @@ On `SCALAR` data are continuously released in batches during defined time interv
 Most of the existing platforms for data science competitions are tailored to offline learning where a static dataset is made available to the participants before the competition starts. This dataset is divided into training and test sets. The training set is used to build and train the model, which is then tested on the test set. 
 
 In online learning, data arrive in a stream of records (instances) and the model needs to be trained incrementally as new instances are released. Since the data arrive at high speed, predictions have to be issued within a short time. Having in mind this specific setup of the data stream mining scenario (Figure \ref{fig:stream_mining}), every model should use a limited 
-amount of time and memory, process one instance at a time and inspect it only once and the model must be able to predict at any time [@bifet2010moa].
+amount of time and memory, process one instance at a time and, inspect it only once and the model must be able to predict at any time [@bifet2010moa].
 
 ![Stream data mining scenario\label{fig:stream_mining}](stream_mining.png)
 
