+++ provider/paper/paper.md
@@ -26,32 +26,31 @@ nocite: '@*'
 
 # Summary
 
-`SCALAR` is a new platform for running real-time machine learning competitions on data streams. Following the intent of Kaggle, which serves as a platform for organizing machine learning competitions adapted for batch learning, we propose `SCALAR` as a novel platform designed specifically for stream learning in real-time. `SCALAR` supports both classification and regression problems in the data streaming setting. It has been developed in `Python`, using state of the art open-source solutions: `Apache Kafka`, `Apache Spark`, `gRPC`, `Protobuf`, and `Docker`. 
+`SCALAR` is a new platform for running real-time machine learning competitions on data streams. Following the intent of Kaggle, which serves as a platform for organizing machine learning competitions adapted for batch learning, we propose `SCALAR` as a novel platform explicitly designed for stream learning in real-time. `SCALAR` supports both classification and regression problems in the data streaming setting. It has been developed in `Python`, using state of the art open-source solutions: `Apache Kafka`, `Apache Spark`, `gRPC`, `Protobuf`, and `Docker`. 
 
 # Statement of need 
 
 Existing machine learning platforms for research and competitions, such as [`Kaggle`](https://www.kaggle.com/) and [`Alibaba Tianchi`](https://tianchi.aliyun.com/) provide static data sets that users leverage to do batch training and scoring. These popular platforms, especially `Kaggle`, enable users to propose better solutions for a wide range of machine-learning problems and applications, pushing forward the whole research community.
 
-While these platforms enable collaborative model building they do not provide real-time settings and data streaming instead of static sets. Inspired by the great reception and adoption that these platforms have had, `SCALAR` was built for real-time settings by supporting data streaming to fill this gap, helping users test and improve their methodologies in a real-time machine learning scenario. 
+While these platforms enable collaborative model building, they do not provide real-time settings and data streaming instead of static sets. Inspired by the great reception and adoption that these platforms have had, `SCALAR` was built for real-time environments by supporting data streaming to fill this gap, helping users test and improve their methodologies in a real-time machine learning scenario. 
 
-On `SCALAR` data are continuously released in batches during defined time intervals, then predictions for each current batch, sent before a defined deadline, are evaluated in real-time, and the results are shown on a live leaderboard. 
+On `SCALAR`, data are continuously released in batches during defined time intervals. Predictions for each current batch sent before a designated deadline are evaluated in real-time, and the results are shown on a live leaderboard.
 
-`SCALAR` can be applied to provide real-time machine learning support in multiple scenarios, one of the more notable and recent ones was being the core organizing and coordination platform for the first  Real-time Machine Learning Competition on Data Streams [@boulegane2019real] as part of the [IEEE Big Data 2019 Cup Challenges](http://bigdataieee.org/BigData2019/BigDataCupChallenges.html), where several approaches of data stream mining were designed, tested, and continue to be improved.
+`SCALAR` can be applied to provide real-time machine learning support in multiple scenarios, one of the more notable and recent ones was its role as the core organizing and coordination platform for the first Real-time Machine Learning Competition on Data Streams [@boulegane2019real] as part of the [IEEE Big Data 2019 Cup Challenges](http://bigdataieee.org/BigData2019/BigDataCupChallenges.html).
 
 
 # Streaming learning setting
 
-Most of the existing platforms for data science competitions are tailored to offline learning where a static dataset is made available to the participants before the competition starts. This dataset is divided into training and test sets. The training set is used to build and train the model, which is then tested on the test set. 
+Most of the existing platforms for data science competitions are tailored to offline learning, where a static dataset is made available to the participants before the competition starts. This dataset is divided into training and test sets. The training set is used to build and train the model, which is then tested on the test set. 
 
[…]
