### 2020-11-30 #2676 @galessiorob
@nedRad88 I added a few more edits on this [PR](https://github.com/nedRad88/SCALAR/pull/10), once those are reviewed and merged, and @atanikan completes their checklist we'll be good to go!
