### 2020-11-30 nedRad88/SCALAR#7 @atanikan
Thanks. Closing the issue.
