### 2020-11-30 #2676 @nedRad88
Thank you all @galessiorob @GregaVrbancic @xiaohk @atanikan For suggestions, comments and reviews.
I closed all issues and merged the last changes.
