### 2020-12-01 #2676 @galessiorob
Thanks so much, @atanikan! 🎉

@nedRad88 The next thing we need from you is to deposit your software in Zenodo and post the DOI here. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing. 🚀
