+++ provider/paper/paper.md
@@ -30,11 +30,11 @@ nocite: '@*'
 
 # Statement of need 
 
-Existing machine learning platforms for research and competitions, such as [`Kaggle`](https://www.kaggle.com/) and [`Alibaba Tianchi`](https://tianchi.aliyun.com/) provide static data sets that users leverage to do batch training and scoring. These popular platforms, especially `Kaggle`, enable users to propose better solutions for a wide range of machine-learning problems and applications, pushing forward the whole research community.
+Existing machine learning platforms for research and competitions, such as [`Kaggle`](https://www.kaggle.com/) and [`Alibaba Tianchi`](https://tianchi.aliyun.com/) provide static data sets that users leverage to do batch training and scoring. These popular platforms, especially `Kaggle`, enable users to propose better solutions for a wide range of machine learning problems and applications, pushing forward the whole research community.
 
-While these platforms enable collaborative model building, they do not provide real-time settings and data streaming instead of static sets. Inspired by the great reception and adoption that these platforms have had, `SCALAR` was built for real-time environments by supporting data streaming to fill this gap, helping users test and improve their methodologies in a real-time machine learning scenario. 
+While these platforms enable collaborative model building, they do not provide real-time settings and data streaming instead of static data sets. Inspired by the great reception and adoption that these platforms have had, `SCALAR` was built for real-time environments by supporting data streaming to fill this gap, helping users test and improve their methodologies in a real-time machine learning scenario. 
 
-On `SCALAR`, data are continuously released in batches during defined time intervals. Predictions for each current batch sent before a designated deadline are evaluated in real-time, and the results are shown on a live leaderboard.
+On `SCALAR`, data are continuously released in batches during defined time intervals. Predictions for each current batch sent before designated deadline are evaluated in real-time, and the results are shown on a live leaderboard.
 
 `SCALAR` can be applied to provide real-time machine learning support in multiple scenarios, one of the more notable and recent ones was its role as the core organizing and coordination platform for the first Real-time Machine Learning Competition on Data Streams [@boulegane2019real] as part of the [IEEE Big Data 2019 Cup Challenges](http://bigdataieee.org/BigData2019/BigDataCupChallenges.html).
 
@@ -47,7 +47,7 @@ In online learning, data arrive in a stream of records (instances), and the mode
 
 ![Stream data mining scenario\label{fig:stream_mining}](stream_mining.png)
 
-The model is being updated using the labeled instances and then evaluated on new unlabeled instances. This scenario represents the prequential evaluation scenario or "test-then-train" setting.
+The model is being evaluated on new unlabeled instances and then updated using the labeled instances. This scenario represents the prequential evaluation scenario or "test-then-train" setting.
 To make this scenario possible in `SCALAR`, we first assume that the records in the data stream arrive in batches and that each batch can be of size `1` or more. Then, we define two different kinds of batches:
 
 * Initial batch: This batch is used to build and train the initial model. It is aimed to avoid the cold start problem and as such, contains both features and targets. The initial batch usually has a large number of instances.
@@ -58,17 +58,17 @@ To make this scenario possible in `SCALAR`, we first assume that the records in 
 
 # Architecture
 
-To support all the aforementioned requirements for stream data mining and to be able to organize competitions in such a scenario, a specifically dedicated platform was needed. To the best of our knowledge, there doesn't exist such a platform. Thus we propose `SCALAR`. Figure \ref{architecture} highlights the architecture of the platform that includes a web application and a streaming engine following the fundamentals of Information Flow Processing (IFP)[@Cugola:12].
[…]
