### 2020-12-05 #2676 @galessiorob
@GregaVrbancic, @atanikan, @xiaohk thank you all so much for lending your time and expertise for this paper!

@nedRad88 a Chief Editor will do the final deposit and after that, your paper will be published on JOSS. Congrats! 🎉
