### 2020-12-05 #2676 @kyleniemeyer
Hello @nedRad88, I'm the editor-in-chief on duty, doing some final checks.

In the Zenodo archive, you have one more person listed as an author than is on the paper. Can you clarify this? Normally we expect/require that the author lists match.
