### 2020-12-05 #2676 @nedRad88
Hi @kyleniemeyer, I edited the zenodo upload. The authors should match now, and the one that was different is now labeled as a contributor. Is that ok now?
